+++ src/Flex/Converter/ClassnameToPath.php
@@ -1,0 +1,12 @@
+<?php
+namespace Flex\Converter;
+
+/**
+ * Class ClassnameToPath
+ *
+ * @author Jeff Tunessen <jeff.tunessen@gmail.com>
+ * @deprecated use NamespaceClassnameToPath
+ */
+class ClassnameToPath extends NamespaceClassnameToPath {
+
+}