+++ src/Flex/Converter/PathToClassname.php
@@ -1,0 +1,12 @@
+<?php
+namespace Flex\Converter;
+
+/**
+ * Class PathToClassname
+ *
+ * @author Jeff Tunessen <jeff.tunessen@gmail.com>
+ * @deprecated use PathToNamespaceClassname
+ */
+class PathToClassname extends PathToNamespaceClassname {
+
+}